separating navigation controller from view controller
