+++ CGMBLEKitUI/TransmitterSettingsViewController.swift
@@ -12,24 +12,6 @@ import LoopKitUI
 import CGMBLEKit
 import ShareClientUI
 
-class TransmitterSettingsNavigationController: SettingsNavigationViewController, GlucoseUnitObserver {
-
-    private var rootViewController: TransmitterSettingsViewController
-
-    init(rootViewController: TransmitterSettingsViewController) {
-        self.rootViewController = rootViewController
-        super.init(rootViewController: rootViewController)
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func glucoseUnitDidChange(to glucoseUnit: HKUnit) {
-        rootViewController.glucoseUnitDidChange(to: glucoseUnit)
-    }
-}
-
 class TransmitterSettingsViewController: UITableViewController {
 
     let cgmManager: TransmitterManager & CGMManagerUI
